refactor: History 내 Count, Filter,List 컴포넌트 분리
