+++ client/src/component/History/Count/Count.js
@@ -1,4 +1,8 @@
-import { getState, subscribeState } from '../../../controller';
+import {
+  getPaymentLength,
+  getState,
+  subscribeState,
+} from '../../../controller';
 import { storeKeys } from '../../../utils/constant';
 
 export class Count {
@@ -20,9 +24,6 @@ export class Count {
 
   render() {
     const history = getState({ key: storeKeys.CURRENT_HISTORY });
-    this.$count = `전체 내역 ${history.reduce(
-      (p, { datas }) => p + datas.length,
-      0,
-    )}건`;
+    this.$count.innerHTML = `전체 내역 ${getPaymentLength()}건`;
   }
 }
